Fix shell-scripts install.sh

+++ objects/shell-scripts/install.sh
@@ -3,5 +3,5 @@
 CONF_DIR=$(realpath `dirname $0`)
 
 mkdir -p $HOME/bin/
-ln -sf $CONF_DIR/scripts $HOME/bin/df
+[ ! -L "$HOME/bin/df" ] && ln -sf $CONF_DIR/scripts $HOME/bin/df
